[Auto] Keep autofix CLI regressions in the integration suite
